Removed tests

+++ eapi/src/heap.c
@@ -72,7 +72,7 @@ void Debug_SicFree(void* ptr) {
 void SicHeapDump() {
 #ifdef _WIN32
 	DWORD total = 0;
-	float ko;
+	double ko;
 	char buf[4096];
 	/* Simply browse the linked list and show the memory block in a (pseudo) fancy format */
 	plink iterator = NULL;
@@ -82,7 +82,7 @@ void SicHeapDump() {
 		sprintf(buf, "Block at %p: %5d byte(s).\n", iterator->ptr, iterator->memory_space);
 		OutputDebugString(buf);
 	}
-	ko = (float)total/1024.0;
+	ko = (double)total/1024.0;
 	sprintf(buf,"TOTAL: %d bytes (%.0f Ko.)\n-----------\n", total, ceil(ko));
 	OutputDebugString(buf);
 #endif
